Fix api to pull from cleaner file
