+++ src/components/Card/Card.css
@@ -23,7 +23,6 @@ h2 {
 
 .headers {
     font-size: 20px;
-    font-weight: 900;
 }
 
 .favorite {
@@ -54,4 +53,8 @@ button {
 
 .fa-meteor {
     color: orangered;
+}
+
+.card-info {
+    font-size: 1.2em;
 }
