+++ src/components/Container/Container.css
@@ -20,3 +20,15 @@
 .back-btn:hover {
     color: yellow;
 }
+
+.showMoreBtn {
+    color: white;
+    background-color: turquoise;
+    margin: 2% 42%;
+    border-radius: 15px;
+    font-family: 'Raleway';
+    font-size: 16px;
+    opacity: 0.8;
+    padding: 10px;
+    width: 150px;
+}
