+++ src/components/Container/Container.jsx
@@ -4,13 +4,14 @@ import './Container.css';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 
-const Container = ({ data, favoriteCard }) => {
+const Container = ({ data, favoriteCard, handleClick, num }) => {
 	const backBtn = (
 		<Link to={'/'} className="back-btn">
 			<i class="fab fa-jedi-order"> Back Home </i>
 		</Link>
 	);
-	const findData = data.map(item => {
+
+	const findData = data.slice(0, num).map(item => {
 		return <Card info={item} key={item.created} favoriteCard={favoriteCard} />;
 	});
 
@@ -18,6 +19,9 @@ const Container = ({ data, favoriteCard }) => {
 		<main className="card-container">
 			{findData}
 			{backBtn}
+			<button className="showMoreBtn" onClick={e => handleClick(e)}>
+				Show More
+			</button>
 		</main>
 	);
 };
